math: Fix UB in lroundl
UBSAN: Undefined behaviour in ../sysdeps/ieee754/float128/../ldbl-128/s_lroundl.c:72:32 left shift of 562949953421312 by 14 cannot be represented in type 'long int'

+++ sysdeps/ieee754/ldbl-128/s_lroundl.c
@@ -69,7 +69,7 @@ __lroundl (_Float128 x)
 	    result = (long int) i0;
 	  else
 	    {
-	      result = ((long int) i0 << (j0 - 48)) | (j >> (112 - j0));
+	      result = (i0 << (j0 - 48)) | (j >> (112 - j0));
 #ifdef FE_INVALID
 	      if (sizeof (long int) == 8
 		  && sign == 1
